requirements.txt
einops==0.8.1
imageio==2.37.0
pandas==2.3.3
pillow==11.3.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
tensorly==0.9.0
torch==2.9.0+cu128
torchaudio==2.9.0+cu128
torchvision==0.24.0+cu128
av==16.0.1
tqdm==4.66.4
matplotlib==3.10.0
imageio[ffmpeg]
flash-attn==2.8.3
timm==1.0.22
opencv-python==4.12.0.88